harden userAgent version checks even more
In nightly and sometimes beta, the version is open ended (e.g. 88+) until I find a feature detection for the next version (which in nightly is impossible). We handle this by always doing two checks. As long as one check passes, then it can't be considered a lie.

The two checks are identical unless the version is open ended: e.g. 88+ could legitimately have 89 in the string
- note: we are only ever one version check behind
- if RFP is off, we just allow the second check to be the next version
- if RFP is on, we only allow the preceding spoofed version, otherwise the checks are identical

A lot of work for RFP just to handle one in every 13 release cycles.