v89
Very little in v89 changes to go on. This is the only bug (of almost 2000) that I could leverage, not covered by a pref: it's a little slow and not exact: at all zoom levels (windows, en-US default fonts)
- the smallest result was approx 0.923 in FF89
- the largest result was  0.65 in FF88 or lower

Not entirely happy, not convinced it's 100% accurate. Will keep digging for others that could compliment or replace this