fixup collation
AFAICT unsupported locales (e.g. made up like "abc" or `undefined`) do not return a consistent hash. They fall back on the app/system locale (not sure which). It definitely doesn't resort to sort().

`javascript.use_us_english_locale` is limited to en-US, and I do not have a non-EN system (en-US, en-CA etc are all the same collation). Regardless of app or system fallback, the unsupported results will not be stable, when we want them to be. So only add each locale as it is supported, by "cheating" and using isVer .. get over it!