TB svg tweaks (#127)
TB requires these items to be hidden (display none) to return 0x0 if missing, otherwise it returns the size of an image placeholder. Which is what we had. But I changed the generic one, about-wordmark.svg, to only be offscreen so I could detect some forks. So now I have included a duplicate element, hidden, and we remeasure if isTB