limit resize event to FF and nonFF desktop
on android blink when scrolling: the inner window metric is updated (the toolbar auto-hides/shows) and this triggers a bypass + lie - because the css-min inner window is NOT updated. since when was scrolling a resize event

So: inner window
- page loads - 980 x 1797 and css-min is 979 x 1797
- scroll so the toolbar disappears, you need to wait a few ms for it to run/render
   - inner window is now 980 x 1730 (so 67 px for the toolbar, yay me!), but css-min hasn't changed

IDK what happens in gecko. I guess I can test this more later, when I use my touch latop