speed up tofu check
the tofu check doesn't affect entropy, it just helps reduce the number of items to check with offset/clientrect/textmetrics and thus saves overall perf if enough tofu exists. But the tofu check itself eats perf. We can speed that up by only checking for obvious tofu, such as when optional OS language packs are not installed or RFP font vis is enabled. This trims it down to half, 21 of 42 = all the tofu in windows RFP (likely similar to mac RFP).