Delete fontlists.html
This no longer makes sense or is worth the upkeep. We can test individual fonts in fontdebug where we can also change the font-weight from the console. TZP2 will automagically alert to missing bundled or unexpected fonts based on TB/FF and RFP font vis + whitelists, so this too is no longer needed