tweak windows font list
remove variants, they need special handling - on a first run (font weight normal) they might not get detected, or they are but fall back to their namesake (e.g. corbel light using corbel - which is a different size). Then on a subsequent load/run those not detected now are, and those that were now use the real font and we get different sizes.

We will test variants a different way, on all OSes