tidy
ac-outputLatency is hardcoded in RFP, otherwise it is slightly variable: new tabs, page reload. So if not RFP green, change the value to "variable" (we don't care about the obvious OS differences, we already know the OS) but display the value on screen. Now we have a stable user audio subsection for compare/testing since 1358149