tweak (#382)
dynamic urlbar, if hidden, alters window.inner - which is used to check RFP is working for screen, available screen and outer. However, viewport element is always the same as window inner, because on android scrollbars are always overlays. So instead for android we will move the element from viewport section to inner section and that becomes our control measurements for RFP. And this is the same value we want to use to determine out dynamic urlbar dimensions. And if dynamic urlbar is hidden we can notate window.inner to explain why it;s different to all the rest of the inner section

For android, we will also not do a viewport section summary, since we're only down to two methods, and they can vary - it's gets funky with pinch to zoom, and reruns and reloads (drag down to reload) and combos of those things

messy'ish code - might refactor and tidy it all up one day