tweak BB heallth checks (#493)
alpha and android are moving to RR - some deterministic checks are not worth tracking or giving false positives to per RR as they change too easily, so tighten these up to only happen on ESR

RFP health checks are not version specific - e.g. I just code them to be the new expected result or fail, so it can be easy/possible/likely that BB on RR will end up triggering some false positives, not sure and not going to fucking check

If it becomes problematic, I can revisit and maybe add a warning if isBB but not ESR but has failed health? IDK, what a mess